+++ .env
@@ -27,9 +27,6 @@ LIVEKIT_API_KEY=APIkgaA2AhYkgMA
 LIVEKIT_API_SECRET=rEVeWO4bFhSim9NCeG8LDA7omQqr1RYuK2ov6TweqWJA
 NEXT_PUBLIC_LIVEKIT_URL=wss://ephemera-hf5x9al1.livekit.cloud
 
-OPENAI_API_KEY=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
-OPENAI_PROJECT_ID=proj_2rQjt0EjWAIL4CcCjrYQiUXZ
-OPENAI_ORGANIZATION_ID=org-BEWxFv5Z3rFwQM43wlYmNDoD
 
 DISCOGS_CONSUMER_KEY=aHTMNdKApdZZZJlSUSgr
 DISCOGS_CONSUMER_SECRET= PZgpydDKAvLglpLgjYKDXRQpoTiXWQfC
