--- lib/openaiclient.ts
@@ -1,7 +1,0 @@
-import OpenAI from "openai";
-
-export const openai = new OpenAI({
-  apiKey: process.env.OPENAI_API_KEY,
-  project: process.env.OPENAI_PROJECT_ID,
-  organization: process.env.OPENAI_ORGANIZATION_ID,
-});
