+++ package.json
@@ -53,7 +53,6 @@
     "nanoid": "^5.0.8",
     "next": "^14.2.30",
     "next-firebase-auth-edge": "^1.6.2",
-    "openai": "^4.87.3",
     "osc": "^2.4.5",
     "pixi.js": "^8.3.4",
     "raw-loader": "^4.0.2",
